refactor(build-support): split ValidationCoordinator IO helpers (N-1)
Extract the caching and locking free functions out of
ValidationCoordinator.swift (744 → 485 lines) to keep the orchestration
pipeline (coordinate → lock → execute → finalize) the clear focus of the
main file. No behavior change — `swift test --filter
InnoDIBuildSupportTests` (82 tests) stays green.

- `ValidationCoordinator+Caching.swift` — shared-run cache IO
  (loadCachedSharedRun, persistResult, persistMetricsArtifact,
  persistSummaryReport, pruneSharedRunDirectories, …).
- `ValidationCoordinator+Locking.swift` — POSIX lock + stale-lock
  recovery (acquireLock, releaseLock, persistLockMetadata,
  recoverStaleLockIfNeeded, lockFileAgeSeconds) plus adjacent IO
  utilities (LockedDataBuffer, validationProcessExists,
  installReadHandler, formatSeconds) and the two LocalizedError
  wrappers that the lock path throws.

Helpers promoted from `private` to `internal` so they remain callable
from the main file.

Part of Phase N — stacked on PR #25.
