+++ Sources/InnoDIBuildSupport/ValidationCoordinator.swift
@@ -483,262 +483,3 @@ package enum ValidationCoordinator {
     }
 }
 
-private func loadCachedSharedRun(
-    resultURL: URL,
-    sharedRunRecordURL: URL
-) -> (result: ValidationCommandResult, record: SharedValidationRunRecord)? {
-    guard
-        let result = loadCachedResult(at: resultURL),
-        let record = loadSharedRunRecord(at: sharedRunRecordURL)
-    else {
-        return nil
-    }
-
-    return (result, record)
-}
-
-private func loadCachedResult(at url: URL) -> ValidationCommandResult? {
-    guard FileManager.default.fileExists(atPath: url.path(percentEncoded: false)) else {
-        return nil
-    }
-
-    guard
-        let data = try? Data(contentsOf: url),
-        let result = try? JSONDecoder().decode(ValidationCommandResult.self, from: data)
-    else {
-        return nil
-    }
-
-    return result
-}
-
-private func persistResult(_ result: ValidationCommandResult, to url: URL) throws {
-    let data = try JSONEncoder().encode(result)
-    try data.write(to: url, options: .atomic)
-}
-
-private func writeStamp(signature: String, result: ValidationCommandResult, to outputDirectoryURL: URL) throws {
-    let stampURL = outputDirectoryURL.appendingPathComponent("dag-validation-stamp.txt")
-    let content = "signature=\(signature)\nexitCode=\(result.exitCode)\n"
-    try content.write(to: stampURL, atomically: true, encoding: .utf8)
-}
-
-private func persistMetricsArtifact(_ artifact: ValidationMetricsArtifact, to url: URL) throws {
-    let data = try JSONEncoder().encode(artifact)
-    try data.write(to: url, options: .atomic)
-}
-
-private func loadSharedRunRecord(at url: URL) -> SharedValidationRunRecord? {
-    guard FileManager.default.fileExists(atPath: url.path(percentEncoded: false)) else {
-        return nil
-    }
-
-    guard
-        let data = try? Data(contentsOf: url),
-        let record = try? JSONDecoder().decode(SharedValidationRunRecord.self, from: data)
-    else {
-        return nil
-    }
-
-    return record
-}
-
-private func persistSharedRunRecord(_ record: SharedValidationRunRecord, to url: URL) throws {
-    let data = try JSONEncoder().encode(record)
-    try data.write(to: url, options: .atomic)
-}
-
-private func persistSummaryReport(_ content: String, to url: URL) throws {
-    try content.write(to: url, atomically: true, encoding: .utf8)
-}
-
-private func acquireLock(at url: URL) throws -> Int32? {
-    let path = url.path(percentEncoded: false)
-    let descriptor = open(path, O_CREAT | O_EXCL | O_RDWR, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH)
-
-    if descriptor >= 0 {
-        return descriptor
-    }
-
-    if errno == EEXIST {
-        return nil
-    }
-
-    throw POSIXLockError(code: errno, path: path)
-}
-
-private func persistLockMetadata(
-    _ metadata: ValidationCoordinatorLockMetadata,
-    descriptor: Int32,
-    path: String
-) throws {
-    let data = try JSONEncoder().encode(metadata)
-    let handle = FileHandle(fileDescriptor: descriptor, closeOnDealloc: false)
-    do {
-        try handle.truncate(atOffset: 0)
-        try handle.seek(toOffset: 0)
-        try handle.write(contentsOf: data)
-        try handle.synchronize()
-    } catch {
-        throw ValidationCoordinatorIOError(path: path, operation: "write lock metadata", underlying: error)
-    }
-}
-
-private func loadLockMetadata(at url: URL) -> ValidationCoordinatorLockMetadata? {
-    guard
-        let data = try? Data(contentsOf: url),
-        let metadata = try? JSONDecoder().decode(ValidationCoordinatorLockMetadata.self, from: data),
-        metadata.pid > 0
-    else {
-        return nil
-    }
-
-    return metadata
-}
-
-private func releaseLock(descriptor: Int32, at url: URL) {
-    close(descriptor)
-    try? FileManager.default.removeItem(at: url)
-}
-
-private func pruneSharedRunDirectories(keepingDirectoryName directoryName: String, in stateDirectoryURL: URL) throws {
-    let fileManager = FileManager.default
-    let entries = try fileManager.contentsOfDirectory(
-        at: stateDirectoryURL,
-        includingPropertiesForKeys: [.isDirectoryKey],
-        options: [.skipsHiddenFiles]
-    )
-
-    for entry in entries {
-        let values = try entry.resourceValues(forKeys: [.isDirectoryKey])
-        guard values.isDirectory == true, entry.lastPathComponent != directoryName else {
-            continue
-        }
-        try? fileManager.removeItem(at: entry)
-    }
-}
-
-private func recoverStaleLockIfNeeded(
-    at url: URL,
-    staleLockAgeSeconds: TimeInterval,
-    runtime: ValidationCoordinatorRuntime
-) throws -> Bool {
-    let fileManager = FileManager.default
-    let path = url.path(percentEncoded: false)
-
-    guard fileManager.fileExists(atPath: path) else {
-        return false
-    }
-
-    let recoveryTokenURL = url.appendingPathExtension("recovering")
-    guard let recoveryDescriptor = try acquireLock(at: recoveryTokenURL) else {
-        return false
-    }
-    defer { releaseLock(descriptor: recoveryDescriptor, at: recoveryTokenURL) }
-
-    guard fileManager.fileExists(atPath: path) else {
-        return false
-    }
-
-    if let metadata = loadLockMetadata(at: url) {
-        guard runtime.processExists(metadata.pid) == false else {
-            return false
-        }
-        runtime.beforeStaleLockRemoval(url)
-        try? fileManager.removeItem(at: url)
-        return fileManager.fileExists(atPath: path) == false
-    }
-
-    guard
-        let ageSeconds = lockFileAgeSeconds(at: path, now: runtime.currentDate()),
-        ageSeconds >= staleLockAgeSeconds
-    else {
-        return false
-    }
-
-    runtime.beforeStaleLockRemoval(url)
-    try? fileManager.removeItem(at: url)
-    return fileManager.fileExists(atPath: path) == false
-}
-
-private func lockFileAgeSeconds(at path: String, now: Date) -> TimeInterval? {
-    guard let attributes = try? FileManager.default.attributesOfItem(atPath: path) else {
-        return nil
-    }
-
-    let referenceDate = (attributes[.modificationDate] as? Date) ?? (attributes[.creationDate] as? Date)
-    guard let referenceDate else {
-        return nil
-    }
-
-    return max(0, now.timeIntervalSince(referenceDate))
-}
-
-private struct POSIXLockError: LocalizedError {
-    let code: Int32
-    let path: String
-
-    var errorDescription: String? {
-        "Failed to acquire validation lock at '\(path)' (errno: \(code))."
-    }
-}
-
-private struct ValidationCoordinatorIOError: LocalizedError {
-    let path: String
-    let operation: String
-    let underlying: Error
-
-    var errorDescription: String? {
-        "Failed to \(operation) at '\(path)': \(underlying.localizedDescription)"
-    }
-}
-
-private final class LockedDataBuffer: @unchecked Sendable {
-    private let lock = NSLock()
-    private var storage = Data()
-
-    var data: Data {
-        lock.lock()
-        defer { lock.unlock() }
-        return storage
-    }
-
-    func append(_ chunk: Data) {
-        guard !chunk.isEmpty else {
-            return
-        }
-
-        lock.lock()
-        storage.append(chunk)
-        lock.unlock()
-    }
-}
-
-private func validationProcessExists(_ pid: Int32) -> Bool {
-    guard pid > 0 else {
-        return false
-    }
-
-    let result = kill(pid, 0)
-    if result == 0 {
-        return true
-    }
-
-    return errno == EPERM
-}
-
-private func formatSeconds(_ value: TimeInterval) -> String {
-    String(format: "%.2f", value)
-}
-
-private func installReadHandler(on handle: FileHandle, buffer: LockedDataBuffer) {
-    handle.readabilityHandler = { readableHandle in
-        let chunk = readableHandle.availableData
-        if chunk.isEmpty {
-            readableHandle.readabilityHandler = nil
-            return
-        }
-
-        buffer.append(chunk)
-    }
-}
